chore: refactor logger
