+++ src/components/http-probe/index.ts
@@ -28,7 +28,7 @@ import { probing } from '../../utils/probing'
 import { validateResponse } from '../../utils/alert'
 import { Probe } from '../../interfaces/probe'
 import { Notification } from '../../interfaces/notification'
-import { probeLog } from '../../utils/logger'
+import { probeLog } from '../logger'
 import { AxiosResponseWithExtraData } from '../../interfaces/request'
 import { log } from '../../utils/pino'
 import { sendAlerts } from '../notification'
