+++ src/index.ts
@@ -29,7 +29,7 @@ import chalk from 'chalk'
 import boxen from 'boxen'
 import { MailData, MailgunData, SMTPData, WebhookData } from './interfaces/data'
 import { looper } from './utils/looper'
-import { printAllLogs } from './utils/logger'
+import { printAllLogs } from './components/logger'
 import { log } from './utils/pino'
 import { closeLog, openLogfile, flushAllLogs } from './utils/history'
 import { notificationChecker } from './components/notification/checker'
